==> notebooks/1_ETL_pipeline.ipynb <==
# Loading library
import os
os.sys.path.append("../")
from scripts.etl_pipeline import *

# Create a Spark Session
spark = (
    SparkSession.builder.appName("ETL Pipeline")
    .config("spark.sql.repl.eagerEval.enabled", True)
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "4g")
    .config("spark.executor.memory", "2g")
    .getOrCreate()
)

# Extract

## BNPL Dataset

As we can't use `urlretrieve` to get the data from Canvas, please download it to your local machine and move it `data/tables`. Then run the code below to unzip the files

# Assign data path
raw_path = "../data"

# Unzip files (Only run once)
"""
for file in os.listdir(f"{raw_path}/tables"):
    if file == ".gitkeep":
        continue
    with zipfile.ZipFile(f"{raw_path}/tables/{file}", "r") as zip_ref:
        zip_ref.extractall(f"{raw_path}/")
    os.remove(f"{raw_path}/tables/{file}")
"""

## External Dataset

Our external dataset will include socio-economic as well as demographics data on different granularity levels such as state and Local Government Area (LGA), collected from the Australia's Bureau of Statistic. Since LGA is a region that includes postcodes and suburbs, we found a dataset that help us map postcode to LGA code. We downloaded the data that maps postcode to LGA code directly as the website doesn't allow us to use `urlretrieve`. The data can be downloaded by clicking [here](https://www.matthewproctor.com/Content/postcodes/australian_postcodes.csv).

We will also use the ABS's **Personal Income in Australia** for each LGA region

INCOME_DATA_URL = "https://www.abs.gov.au/statistics/labour/earnings-and-working-conditions/personal-income-australia/2020-21-financial-year/Table%202%20-%20Total%20income%20distribution%20by%20geography%2C%202020-21.xlsx"

urlretrieve(INCOME_DATA_URL, f"{raw_path}/raw/income_data.xlsx")

Getting the ABS's **Personal Fraud** statistics, which includes card fraud, identity theft, and scams.

PERSONAL_FRAUD_DATA_URL = "https://www.abs.gov.au/statistics/people/crime-and-justice/personal-fraud/2022-23/Personal%20Fraud%20%28Tables%201a%20to%2014b%29.xlsx"

urlretrieve(PERSONAL_FRAUD_DATA_URL, f"{raw_path}/raw/personal_fraud.xlsx")

# Transform

## External Dataset

For the data on mapping postcodes to LGA code, we will only select the columns that are neccessary for the mapping. There are some entries with missing LGA code. To resolve this, we will use the closest postcode, using lattitude and longitude, and if it has a valid LGA code, we will impute the missing value with the existing one. To do this, we will using K-Nearest Neighour wth `k=1` to do. The reason being the data is already being sorted (clustered) using states. Thus, we only need to select the closest postcode with a valid LGA to the target.


postcode_lga_map = pd.read_csv("../data/raw/australian_postcodes.csv")

# Clean the data
postcode_lga_map = clean_postcode_lga_mapping(postcode_lga_map)

# Display the data
postcode_lga_map

For the data on median/mean income and age of earners, as the data is in excel format, we did some extensive work to renaming the columns as well as selecting the correct rows that contain the number we want.

There are LGA codes in Western Australia that do not have any values across all features that we're planning to use. Thus, we will impute the median age, income and mean income using the state's values. 

For the number of earners, we notice that there's a discrepancy in the total number of earners. In total, the number of earners in Western Australia is 1,585,093 whereas when summing up the number of earners across all LGA code in Western Australia, the number is 1,5815,061. This is a difference of 32 earners. We're not entirely sure why this discrepancy occurs but we will split this number in half for the 2 missing LGA code. Even though this is not a sound solution, we believe that imputing such number for only 2 entries doesn't affect the fraud probability predition significantly as well as we're not entirely sure if the consumers live in this LGA code.

income_df = pd.read_excel("../data/raw/income_data.xlsx", sheet_name = "Table 2.5", skiprows = 5, skipfooter = 5)

income_df = preprocess_income_df(income_df)
income_df.head(10)

Merging the income data with the postcode data.

print(f"Shape of postcode data before the merge: {postcode_lga_map.shape}")
postcode_info = postcode_lga_map.merge(income_df, left_on='lgacode', right_on='lga',
                                       how = "left" )
print(f"Shape of postcode data after the merge: {postcode_info.shape}")

We also need to check that if there are any postcode with LGA code that is not in the income data

postcode_info[postcode_info.lga.isnull()]

We can see that there are 2 postcode that does not appear in the income data. We will follow the same approach, using K-Nearest Neighbour, that we did previously to compute these value.

postcode_info = impute_income_metrics(postcode_info)
postcode_info = postcode_info.drop(columns='lga')

The ABS's **Personal Fraud** data have multiple tables, in which we will use Table 4a and Table 4b as it contains the percentage of individual experienced personal fraud and respective Relative Standard Error (RSE) for each state, respectively. We will use the rate from 2021 to 2022 to match our the date range of the data we have one the transactions.

pf_df = process_fp_data(path="../data/raw/personal_fraud.xlsx")
pf_df

## Main Dataset

The system use `user_id` as a key for identifying customer in transactions record and fraud probability tables. However, they also have a key-value map of `user_id` and `consumer_id`. We will use `consumer_id` as the only ID for customer. Thus, we will map `user_id` from each table to `consumer_id` and drop the former.

# Load consumer user details -> a key:value map for user_id to consumer_id
consumer_user_map = spark.read.parquet(f"{raw_path}/tables/consumer_user_details.parquet")
consumer_user_map.limit(5) # Preview

# Load consumer fraud rate dataset
consumer_fraud_rate = spark.read.csv(f"{raw_path}/tables/consumer_fraud_probability.csv", header=True, inferSchema=True)
consumer_fraud_rate.show(5)
get_dataset_count(consumer_fraud_rate)

# Replace all user_id with unique consumer_id
consumer_fraud_rate = replace_id(consumer_user_map, consumer_fraud_rate)
consumer_fraud_rate.show(5)

# Check to make sure no rows were lost on the inner join
get_dataset_count(consumer_fraud_rate)

We see that there is no change in the number of entries upon an inner join.

consumer_fraud_rate.groupBy('consumer_id').agg(F.avg("fraud_probability").alias('avg_fp')).count()

# Load all the transaction data 
transaction_p1 = spark.read.parquet(f"{raw_path}/tables/transactions_20210228_20210827_snapshot")
transaction_p2 = spark.read.parquet(f"{raw_path}/tables/transactions_20210828_20220227_snapshot")
transaction_p3 = spark.read.parquet(f"{raw_path}/tables/transactions_20220228_20220828_snapshot")

# Combine the datasets
transaction_records = reduce(DataFrame.unionAll, [transaction_p1, transaction_p2, transaction_p3])
get_dataset_count(transaction_records)

# Replace user_id with consumer_id after combining
transaction_records = replace_id(consumer_user_map, transaction_records)

# Check to make sure no rows were lost on the inner join
get_dataset_count(transaction_records)

## Cleaning

Now that replacing `user_id` to `consumer_id` is done, we will load all other data and clean them altogether. We start off with dropping duplicates from the consumer fraud probability data

print(f"Number of entries before dropping duplicates: {consumer_fraud_rate.count()}")
consumer_fraud_rate = consumer_fraud_rate.dropDuplicates()
print(f"Number of entries after dropping duplicates: {consumer_fraud_rate.count()}")

We can see that around 0.28% of the entries are duplicates.

Doing the same thing for the transactions data.

print(f"Number of entries before dropping duplicates: {transaction_records.count()}")
transaction_records = transaction_records.dropDuplicates()
print(f"Number of entries after dropping duplicates: {transaction_records.count()}")

There are no duplicates in the transactions data. Next, we will load the merchant fraud rate and repeat the same process.

# Load consumer fraud rate dataset
merchant_fraud_rate = spark.read.csv(f"{raw_path}/tables/merchant_fraud_probability.csv", header=True, inferSchema=True)
merchant_fraud_rate.show(5)
get_dataset_count(merchant_fraud_rate)

Removing any duplicates existing in the data

print(f"Number of entries before dropping duplicates: {merchant_fraud_rate.count()}")
merchant_fraud_rate = merchant_fraud_rate.dropDuplicates()
print(f"Number of entries after dropping duplicates: {merchant_fraud_rate.count()}")

Cleaning `tbl_merchants.parquet`. The feature `tags` is a string that represents either a tuple or a list, containing 3 elements:
* Items that are being sold
* Revenue levels
* Commission rate

Each elements either a list, a tuple, or a combination of both (e.g starts with `[` and ends with `)` and vice versa). These inconsistencies are mostly due to human errors. Thus, we need to take into account these consistent when splitting the values of the feature `tags` into separate columns

# Load merchant's info
merchant_info = spark.read.parquet(f"{raw_path}/tables/tbl_merchants.parquet")
merchant_info.show(5, truncate=False)

# Clean the data
merchant_info = clean_merchant_details(merchant_info)
merchant_info.show(5, truncate=False)

Removing any duplicated merchants.

print(f"Number of entries before dropping duplicates: {merchant_info.count()}")
merchant_info = merchant_info.dropDuplicates()
print(f"Number of entries after dropping duplicates: {merchant_info.count()}")

The data on consumer's basic information is a single column that contains the consumer's name, address, state, postcode, gender, and their unqiue consumer ID, each separated by "`|`". Thus, we will need to split these into individual columns. Based on the `README.md` for the data, we will only keep the consumer's name, state, postcode, gender, and consumer ID as the addresses are fake.

# Load consumer info's
consumer_info = spark.read.csv(f"{raw_path}/tables/tbl_consumer.csv", header=True, inferSchema=True)
consumer_info.show(5, truncate=False)

# Clean the data
consumer_info = clean_consumer_details(consumer_info)
consumer_info.show(5, truncate=False)

Removing any duplicated consumers.

print(f"Number of entries before dropping duplicates: {consumer_info.count()}")
consumer_info = consumer_info.dropDuplicates()
print(f"Number of entries after dropping duplicates: {consumer_info.count()}")

For all data (transactions, merchant and consumer fraud probability), we will need to ensure that the datetime of all dataset with such column is within the specified range (labeled on the name of the intial downloaded file)

transaction_records = ensure_datetime_range(transaction_records, "2021-03-01", "2022-08-28")

merchant_fraud_rate = ensure_datetime_range(merchant_fraud_rate, "2021-03-01", "2022-08-28")

consumer_fraud_rate = ensure_datetime_range(consumer_fraud_rate, "2021-03-01", "2022-08-28")

Next, we check for any existing null values across all dataset.

calculate_missing_values(consumer_fraud_rate)
calculate_missing_values(merchant_fraud_rate)
calculate_missing_values(consumer_info)
calculate_missing_values(merchant_info)
calculate_missing_values(transaction_records)

Currently, there are no mising values after we do some cleaning. We will come back to this after we merged the data together.

# Load

landing_directory = "../data/curated"

# Main dataset
consumer_fraud_rate.write.parquet(f"{landing_directory}/consumer_fp.parquet", mode = 'overwrite')
merchant_fraud_rate.write.parquet(f"{landing_directory}/merchant_fp.parquet", mode = 'overwrite')
transaction_records.write.parquet(f"{landing_directory}/transactions.parquet", mode = 'overwrite')
merchant_info.write.parquet(f"{landing_directory}/merchant_info.parquet", mode = 'overwrite')
consumer_info.write.parquet(f"{landing_directory}/consumer_info.parquet", mode = 'overwrite')

# External dataset
postcode_info.to_csv("../data/curated/postcode_info.csv")
pf_df.to_csv("../data/curated/personal_fraud.csv")
