==> notebooks/2.2_geospatial_analysis.ipynb <==
import os
os.sys.path.append("../")
from scripts.geospatial_analysis import *

# Create a Spark Session
spark = (
    SparkSession.builder.appName("geospatial analysis")
    .config("spark.sql.repl.eagerEval.enabled", True)
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "4g")
    .config("spark.execturo.memory", "2g")
    .getOrCreate()
)

consumer_info = spark.read.parquet('../data/curated/consumer_info.parquet')
fraudulent_consumer_rate = spark.read.parquet('../data/curated/consumer_fraud_prob.parquet')

fraudulent_consumer_with_info = consumer_info.join(fraudulent_consumer_rate, on="consumer_id", how="inner")

# Find info for fraudulent consumer
fraudulent_consumer_with_info = consumer_info.join(fraudulent_consumer_rate, on="consumer_id", how="inner")

# Number of consumer in each state
consumer_info.groupby("state").agg(F.count("consumer_id"))

# Number of fraudulent consumer in each state
fraudulent_consumer_with_info.groupby("state").agg(F.count("consumer_id"))

fraudulent_consumer_with_info.count()
#consumer_info.count()
#fraudulent_consumer_with_info.select("consumer_id").distinct().count() # Number of unique fraudulent consumer

# Check the number of NULL values
null_counts = fraudulent_consumer_with_info.select([F.count(F.when(F.col(c).isNull(), c)).alias(c) for c in fraudulent_consumer_with_info.columns])
null_counts.limit(1)
# null_counts = consumer_info.select([F.count(F.when(F.col(c).isNull(), c)).alias(c) for c in fraudulent_consumer_with_info.columns])
# null_counts.limit(1)


# Number of consumer in each postcode
consumer_info.groupBy(["postcode"]).agg(F.count("consumer_id").alias("total_consumer"))

# External dataset

ABS Digital Boundary files

`States and Territories - 2021 - Shapefile`

`Australian postcodes - Shapefile`

https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files#downloads-for-gda2020-digital-boundary-files


# Transform the full state name to the abbriviation name
state_abbr = {
    "New South Wales": "NSW",
    "Victoria": "VIC",
    "Queensland": "QLD",
    "South Australia": "SA",
    "Western Australia": "WA",
    "Tasmania": "TAS",
    "Northern Territory": "NT",
    "Australian Capital Territory": "ACT"
}

# Load the shapefile into a GeoDataFrame
postcode_shapefile_path = '../data/raw/POA_2021_AUST_GDA2020_SHP/POA_2021_AUST_GDA2020.shp'
gdf_postcode = gpd.read_file(postcode_shapefile_path)

state_shapefile_path = '../data/raw/STE_2021_AUST_SHP_GDA2020/STE_2021_AUST_GDA2020.shp'
gdf_state = gpd.read_file(state_shapefile_path)
gdf_state['STE_NAME21'] = gdf_state['STE_NAME21'].replace(state_abbr)

fraudulent_consumer_with_info

# Number of (fraudulent) consumer in each postcode or state
consumer_group_by_postcode = consumer_info.groupBy(["postcode"]).agg(F.count("consumer_id").alias("total_consumer"))
consumer_group_by_state = consumer_info.groupBy(["state"]).agg(F.count("consumer_id").alias("total_consumer"))

fraudulent_consumer_group_by_postcode = fraudulent_consumer_with_info.groupBy(["postcode"]).agg(F.count("consumer_id").alias("total_consumer"),
                                                                                                 F.avg("fraud_probability").alias("average_fraud_prob"),
                                                                                                 F.max("fraud_probability").alias("max_fraud_probability"))
fraudulent_consumer_group_by_state = fraudulent_consumer_with_info.groupBy(["state"]).agg(F.count("consumer_id").alias("total_consumer"),
                                                                                                 F.avg("fraud_probability").alias("average_fraud_prob"),
                                                                                                 F.max("fraud_probability").alias("max_fraud_probability"))

# Group the consumer and fraudulent consumer by different features
consumer_group_by_postcode = consumer_group_by_postcode.toPandas()
consumer_group_by_state = consumer_group_by_state.toPandas()

fraudulent_consumer_group_by_postcode = fraudulent_consumer_group_by_postcode.toPandas()
fraudulent_consumer_group_by_state = fraudulent_consumer_group_by_state.toPandas()

consumer_group_by_postcode['postcode'] = consumer_group_by_postcode['postcode'].astype(str)
consumer_group_by_state['state'] = consumer_group_by_state['state'].astype(str)

fraudulent_consumer_group_by_postcode['postcode'] = fraudulent_consumer_group_by_postcode['postcode'].astype(str)
fraudulent_consumer_group_by_state['state'] = fraudulent_consumer_group_by_state['state'].astype(str)

gdf_postcode['POA_CODE21'] = gdf_postcode['POA_CODE21'].astype(str)

# Merge consumer information with geographic data from the shapefile
total_consumer_group_by_postcode_merged_gdf = gdf_postcode.merge(consumer_group_by_postcode, left_on='POA_CODE21', right_on='postcode')
total_consumer_group_by_postcode__geojson = total_consumer_group_by_postcode_merged_gdf.to_json()

fraudulent_consumer_group_by_postcode_merged_gdf = gdf_postcode.merge(fraudulent_consumer_group_by_postcode, left_on='POA_CODE21', right_on='postcode')
fraudulent_consumer_group_by_postcode_geojson = fraudulent_consumer_group_by_postcode_merged_gdf.to_json()

# Merge consumer information with geographic data from the shapefile
total_consumer_group_by_state_merged_gdf = gdf_state.merge(consumer_group_by_state, left_on='STE_NAME21', right_on='state')
total_consumer_group_by_state__geojson = total_consumer_group_by_state_merged_gdf.to_json()

fraudulent_consumer_group_by_state_merged_gdf = gdf_state.merge(fraudulent_consumer_group_by_state, left_on='STE_NAME21', right_on='state')
fraudulent_consumer_group_by_state_geojson = fraudulent_consumer_group_by_state_merged_gdf.to_json()

# Distribution of number of customer group by postcode
# Computer might be lagging, better to clear output of this cell
create_consumer_map(total_consumer_group_by_postcode__geojson,consumer_group_by_postcode, "postcode", "total_consumer", "POA_CODE21")

# Distribution of average fraud probability group by postcode
create_consumer_map(fraudulent_consumer_group_by_postcode_geojson,fraudulent_consumer_group_by_postcode, "postcode", "average_fraud_prob", "POA_CODE21")

# Distribution of maximum fraud probability group by postcode
create_consumer_map(fraudulent_consumer_group_by_postcode_geojson,fraudulent_consumer_group_by_postcode, "postcode", "max_fraud_probability", "POA_CODE21")

# Group by state instead
create_consumer_map(total_consumer_group_by_state__geojson,consumer_group_by_state, "state", "total_consumer","STE_NAME21")

create_consumer_map(fraudulent_consumer_group_by_state_geojson,fraudulent_consumer_group_by_state, "state", "average_fraud_prob","STE_NAME21")

create_consumer_map(fraudulent_consumer_group_by_state_geojson,fraudulent_consumer_group_by_state, "state", "max_fraud_probability","STE_NAME21")
