==> notebooks/3.2_merchant_fraud_model.ipynb <==
import os
os.sys.path.append("../")
from scripts.merchant_fraud import *


spark = (
    SparkSession.builder.appName("Merchant Fraud Model")
    .config("spark.sql.repl.eagerEval.enabled", True)
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.driver.memory", "4g")
    .config("spark.executor.memory", "2g")
    .getOrCreate()
)

First, let's load the data that we're going to use

path = "../data/curated"

# Read in transactions dataset
transactions = spark.read.parquet(f"{path}/transactions.parquet")

## Feature engineering

We decide to introduce feature that will flag a transaction with a dollar value that significantly deviates from the mean dollar value. We going to keep it simple and assume that the underlying distribution of the dollar value for all merchants' transactions is normal. if the error, difference between dollar value and average dollar value, is greater than 2 standard deviation then the transaction will be flagged as suspicious.

We recognise that this may not be the best practice as the underlying transaction's dollar value distribution is normal for all merchants. Regardless, we believe this simple approach suffices for flagging transactions.

# Flag unusual transactions that deviate greatly from a merchant's usual dollar value
# Calculate average and standard deviation of dollar_value per merchant
transaction_stats = transactions.groupBy("merchant_abn").agg(
    F.avg("dollar_value").alias("avg_dollar_value"),
    F.stddev("dollar_value").alias("std_dollar_value")
)

# Join the stats back to the original dataset
transaction_df = transactions.join(transaction_stats, on="merchant_abn", how="left")

We noticed that there are certain merchant with `NULL` values for dollar value standard deviation. This is because that these merchants only have one transaction across the entirety time range of the data. Though these records seems suspcious as it seems unreasonable for a business to have only 1 transaction from Febuary 2021 to August 2022, we still going to keep it in the data.

# Calculate how many standard deviations away each transaction is, in other words, we're normalising the dollar value
# May need extra caution to interpret this feature as it can be POSITIVE OR NEGATIVE
transaction_df = transaction_df.withColumn(
    "std_diff_dollar_value", 
    F.when(
        F.col("std_dollar_value").isNotNull() & (F.col("std_dollar_value") != 0),  # Are there cases where std of dollar value is 0?
        (F.col("dollar_value") - F.col("avg_dollar_value")) / F.col("std_dollar_value")
    ).otherwise(0) 
)
transaction_df = transaction_df.drop("consumer_id", "consumer_fp", "name", "category", "avg_dollar_value", "std_dollar_value")

# Flag unusual monthly transaction volumes that deviate from a merchant's usual monthly volume
# Extract month and year from order_datetime
transaction_df = transaction_df.withColumn("order_month", F.date_format(F.col("order_datetime"), "yyyy-MM"))

# Calculate number of transactions per merchant per month
transaction_records_monthly = transaction_df.groupBy("merchant_abn", "order_month").agg(
    F.count("order_id").alias("monthly_order_volume")
)

# Calculate the average and standard deviation of monthly transactions per merchant
transaction_stats = transaction_records_monthly.groupBy("merchant_abn").agg(
    F.avg("monthly_order_volume").alias("avg_monthly_order_volume"),
    F.stddev("monthly_order_volume").alias("stddev_monthly_order_volume")
)

# Join the monthly volume feature back with the original dataset
transaction_records_final = transaction_df.join(transaction_records_monthly, on=["merchant_abn", "order_month"], how="left"
)

# Join the transaction statistics back to the original dataset 
transaction_records_final = transaction_records_final.join(transaction_stats, on="merchant_abn", how="left")

# Calculate how many standard deviations away each monthly volume is
transaction_records_final = transaction_records_final.withColumn(
    "std_diff_order_volume", 
    F.when(F.col("stddev_monthly_order_volume").isNotNull() & (F.col("stddev_monthly_order_volume") != 0),
           (F.col("monthly_order_volume") - F.col("avg_monthly_order_volume")) / F.col("stddev_monthly_order_volume"))
    .otherwise(0)
)

transaction_records_final.show(5)

transaction_records_final = transaction_records_final.drop("avg_monthly_order_volume", "stddev_monthly_order_volume")

We suspect that there is an underlying relationship between month, day of the week, and if the order date is a weekend with the fraud probability of the merchants. Thus, we will create 3 features that capture this temporal effect. We will later encode these feature to feed into our model

# Extract the weekday (1 = Sunday, 7 = Saturday)
transaction_records_final = transaction_records_final.withColumn("weekday", F.dayofweek("order_datetime"))

# Add a column to flag weekends (Saturday = 7, Sunday = 1)
transaction_records_final = transaction_records_final.withColumn(
    "is_weekend", 
    F.when((F.col("weekday") == 7) | (F.col("weekday") == 1), 1).otherwise(0)
)

# Extract year and month from 'order_month' and create new columns
transaction_records_final = transaction_records_final.withColumns(
    {"year":  F.split(F.col("order_month"), "-")[0].cast("integer"),
    "month": F.split(F.col("order_month"), "-")[1].cast("integer")}
)

# Modelling

Splitting the data to train and test set to train and fine-tune our model.

train_data = transaction_records_final.filter(F.col("merchant_fp").isNotNull())
test_data = transaction_records_final.filter(F.col("merchant_fp").isNull())

We will be using Random Forest Regression and Linear Regression to predict the fraud probability. We will aso perform a cross-validated grid search to find the "better" hyperparameters for the model.

For now, we will use the default hyperparameters of the model to see how each model perform.

# Assemble the data
assembled_train, _ = assemble_data(train_data)
train_set, validation_set = assembled_train.randomSplit([0.8,0.2], seed=123)

rfr_model = unoptimal_model(RandomForestRegressor(labelCol='merchant_fp', featuresCol='features'),
                            train_set, validation_set)

# lr_model = unoptimal_model(LinearRegression(labelCol="merchant_fp", featuresCol="features"),
#                            train_set, validation_set)

We can see that the LR performance is quite terrible with RMSE that's almost double than that of RFR. It's $R^2$ is only 0.08 which indicates that the model failed to explain a large portion of variation in the data. Thus, we will use RFR as our main model and perform cross-validated grid search.

# Parameter grid
rfr_paramGrid = ParamGridBuilder() \
    .addGrid(RandomForestRegressor(labelCol='merchant_fp', featuresCol='features').numTrees, [10, 20, 40]) \
    .addGrid(RandomForestRegressor(labelCol='merchant_fp', featuresCol='features').maxDepth, [5, 10, 12]) \
    .build()

evaluator = RegressionEvaluator(labelCol="merchant_fp", predictionCol="prediction")

crossval = CrossValidator(estimator=RandomForestRegressor(labelCol='merchant_fp', featuresCol='features'),
                          estimatorParamMaps=rfr_paramGrid,
                          evaluator=evaluator,
                          numFolds=2)

cv_model = crossval.fit(train_set)
cv_predictions = cv_model.transform(validation_set)
cv_rmse = evaluator.evaluate(cv_predictions)

best_model = cv_model.bestModel
print(f"Best Model RMSE on test data = {cv_rmse}")
print(f"Best number of trees: {best_model.getNumTrees}") 
print(f"Best max depth: {best_model.getMaxDepth()}") 
print(f"Best max bins: {best_model.getMaxBins()}")

rfr_better = RandomForestRegressor(labelCol='merchant_fp', featuresCol='features',
                                         numTrees = 20, maxBins = 32, maxDepth=5)

rfr_better_model = rfr_better.fit(train_set)

assembled_test, assembler = assemble_data(test_data)

Making predictions on the test data and combine it with the existing probabilities.

predictions = rfr_better_model.transform(assembled_test)
predictions = predictions.select(*['merchant_abn', 'order_datetime', 'order_id', 'prediction'])
predictions = predictions.withColumnRenamed("prediction", "merchant_fp")

train_data = train_data.select(*['merchant_abn', 'order_datetime', 'order_id', 'merchant_fp'])

final_df = train_data.union(predictions)
# predictions.write.parquet(f"../data/curated/transactions_predicted_merchant_fp.parquet", mode = "overwrite")

final_df.write.parquet(f"../data/curated/predicted_merchant_fp.parquet")

## Feature importances

import pandas as pd

rfr_feature_importances = rfr_better_model.featureImportances
feature_names = assembler.getInputCols()

rf_importances_df = pd.DataFrame({
    "Feature": feature_names,
    "Importance": rfr_feature_importances.toArray()
}).sort_values(by="Importance", ascending=False)

print(rf_importances_df)
